week_7_project/requirements.txt
pandas==1.5.2
prefect-gcp[cloud_storage]
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
nltk==3.7
praw==7.7.0
spacy==3.3.1
python-decouple